Worm game in development
This is the first iteration of the worm game. So far there is a window with a white background, the snake which moves on its own, and some text on the screen. Passing control of the worm to the player still needs to be implemented. I also started the code on the food the worm eats.