Worm Game in Development
Added a score counter and worm food. Any time the worm touches the food the score increases and a new food spawns randomly. The exponential movement bug from the last patch was fixed. Handling the worm touching the boundaries is still being worked on. A known bug right now is if the worm touches a arbitrary boundary it will no longer accept movement input. The fruit may also spawn out of view of the game window.