Worm Game Complete (Features To Be Added Later)
The boundary bugs are fixed for both the worm and worm food. This version marks the game as playable with no bugs. There are still some features that can be added such changing movement speed and object size, but those have no timeline to be added until the next game is made. The next necessary change for this game is to add support to be managed by the Arcade Cabinet.