Paddle Game in Development
Currently playable. The player controls the left paddle with the up and down arrow keys. The enemy paddle and ball move on their own. A score counter is implemented and working. The known bug is how the ball interacts with the top and bottom boundary of the game. It may go off screen and/or act erratically near these borders.