Fixed some minor errors
Fixed some incorrect naming in comments. Fixed a bug that stopped the ball from resetting after scoring. Add a delay when scoring so the player knows the ball is resetting. The enemy paddle's width is now showing correctly (both paddles should look identical). The hit detection was also adjusted. It is not perfect but should be more accurate.