Start of new game
This is the start of a space invader like game called space game. It uses the worm game as starting code so until this game is completed it will have have pieces of worm game spread apparent. The other new file SpaceObject.java will be needed to run this in the future.