Integrated SpaceObject.java
Added integration with SpaceObject.java. Added in objects for each thing shown on screen. Removed most of the unneeded worm game code. The stuff left in will be useful to adapt.  Also changed the colors shown in the game to be more consistent with the desired view of the game.