Fixed bug when selecting same game
Fixed a bug that allowed multiple instances of the same game to be open at once if you hit submit while playing the same game as selected. Added instructions to alert player that they have to click in a game so the game can use the keyboard inputs. Simplified the help menu.