Added pause button and new fire button
Based on customer feedback a pause button was added. The player can now use the space button to fire instead of the up arrow if desired. The enemy fire rate was adjusted to avoid targeting barriers too often. The enemy bullets no longer disappear immediately after an enemy is destroyed. This increases the game difficulty.